Remove Wrapper component and pass functionality to OrgView
