+++ src/App.js
@@ -14,7 +14,6 @@ import {
   parsePath,
 } from './routing'
 import initWrapper, { pollConnectivity } from './aragonjs-wrapper'
-import Wrapper from './Wrapper'
 import { Onboarding } from './onboarding'
 import { getWeb3 } from './web3-utils'
 import { log } from './utils'
@@ -27,6 +26,7 @@ import LocalIdentityModal from './components/LocalIdentityModal/LocalIdentityMod
 import HelpScoutBeacon from './components/HelpScoutBeacon/HelpScoutBeacon'
 import GlobalPreferences from './components/GlobalPreferences/GlobalPreferences'
 import CustomToast from './components/CustomToast/CustomToast'
+import OrgView from './components/OrgView/OrgView'
 
 import { isKnownRepo } from './repo-utils'
 import {
@@ -352,7 +352,6 @@ class App extends React.Component {
       appIdentifiers,
       appsStatus,
       canUpgradeOrg,
-      connected,
       daoAddress,
       daoStatus,
       fatalError,
@@ -430,19 +429,17 @@ class App extends React.Component {
                           permissions={permissions}
                         >
                           <div css="position: relative; z-index: 0">
-                            <Wrapper
+                            <OrgView
                               visible={mode === APP_MODE_ORG}
                               apps={appsWithIdentifiers}
                               appsStatus={appsStatus}
                               canUpgradeOrg={canUpgradeOrg}
-                              connected={connected}
                               daoAddress={daoAddress}
                               daoStatus={daoStatus}
                               historyBack={this.historyBack}
                               historyPush={this.historyPush}
                               locator={locator}
-                              onRequestAppsReload={this.handleRequestAppsReload}
-                              openPreferences={this.openPreferences}
+                              onOpenPreferences={this.openPreferences}
                               permissionsLoading={permissionsLoading}
                               repos={repos}
                               signatureBag={signatureBag}
